Improvements and addition of the British flag.
Bugfix: Changes to cross and cross2 methods so that flags that
have crosses that are aligned to the left will not appear
skewed for most aspect ratios.

Added centeredCross and britishDiagonalCross methods so that
the UK-flag is supported.

Changed the api from 'new Flag()' to 'flags()' with the reasoning
that it is in fact a utility function does DOM scripting not an
artifact that needs to be instantiated.